Consolidate shared utils
